refactor(resolver): split import resolution into language-specific submodules
